+++ djstripe/models/__init__.py
@@ -3,14 +3,17 @@ from .billing import (
 	Coupon, Invoice, InvoiceItem, Plan, Subscription,
 	SubscriptionItem, UpcomingInvoice, UsageRecord
 )
+from .checkout import Session
 from .connect import (
 	Account, ApplicationFee, ApplicationFeeRefund, CountrySpec, Transfer, TransferReversal
 )
 from .core import (
-	BalanceTransaction, Charge, Customer, Dispute,
-	Event, FileUpload, Payout, Product, Refund
+	BalanceTransaction, Charge, Customer, Dispute, Event,
+	FileUpload, PaymentIntent, Payout, Product, Refund, SetupIntent
 )
-from .payment_methods import BankAccount, Card, DjstripePaymentMethod, Source
+from .payment_methods import (
+	BankAccount, Card, DjstripePaymentMethod, PaymentMethod, Source
+)
 from .sigma import ScheduledQueryRun
 from .webhooks import WebhookEventTrigger
 
@@ -32,10 +35,14 @@ __all__ = [
 	"IdempotencyKey",
 	"Invoice",
 	"InvoiceItem",
+	"PaymentIntent",
+	"PaymentMethod",
 	"Payout",
 	"Plan",
 	"Product",
 	"Refund",
+	"SetupIntent",
+	"Session",
 	"ScheduledQueryRun",
 	"Source",
 	"StripeModel",