+++ tests/apps/example/forms.py
@@ -10,3 +10,7 @@ class PurchaseSubscriptionForm(forms.Form):
 	stripe_source = forms.CharField(
 		max_length="255", widget=forms.HiddenInput(), required=False
 	)
+
+
+class PaymentIntentForm(forms.Form):
+	pass