+++ tests/apps/example/management/commands/regenerate_test_fixtures.py
@@ -92,6 +92,8 @@ class Command(BaseCommand):
 				"last4",
 				"tokenization_method",
 			],
+			djstripe.models.PaymentIntent: ["id"],
+			djstripe.models.PaymentMethod: ["id"],
 			djstripe.models.Source: [
 				"id",
 				"amount",
@@ -125,11 +127,10 @@ class Command(BaseCommand):
 		Fields that we don't care about the value of, and that preserving
 		allows us to avoid churn in the fixtures
 		"""
-		# TODO payment_intent from this once PaymentIntent model / fixture exists
 		model_sideeffect_fields = {
 			djstripe.models.BalanceTransaction: ["available_on"],
 			djstripe.models.Source: ["client_secret"],
-			djstripe.models.Charge: ["payment_intent", "receipt_url"],
+			djstripe.models.Charge: ["receipt_url"],
 			djstripe.models.Subscription: [
 				"billing_cycle_anchor",
 				"current_period_start",
@@ -150,7 +151,6 @@ class Command(BaseCommand):
 				"webhooks_delivered_at",
 				"period_start",
 				"period_end",
-				"payment_intent",
 				# we don't currently track separate fixtures for SubscriptionItems
 				"subscription_item",
 			],
@@ -180,6 +180,8 @@ class Command(BaseCommand):
 			],
 			djstripe.models.Invoice: [tests.FAKE_INVOICE],
 			djstripe.models.Charge: [tests.FAKE_CHARGE],
+			djstripe.models.PaymentIntent: [tests.FAKE_PAYMENT_INTENT_I],
+			djstripe.models.PaymentMethod: [tests.FAKE_PAYMENT_METHOD_I],
 			djstripe.models.BalanceTransaction: [tests.FAKE_BALANCE_TRANSACTION],
 		}
 
@@ -261,6 +263,11 @@ class Command(BaseCommand):
 			# assume that test customers don't have more than 100 cards...
 			for card in customer.sources.list(limit=100):
 				self.update_fake_id_map(card)
+
+			for payment_method in djstripe.models.PaymentMethod.api_list(
+				customer=customer.id, type="card"
+			):
+				self.update_fake_id_map(payment_method)
 
 			for subscription in customer["subscriptions"]["data"]:
 				self.update_fake_id_map(subscription)
@@ -360,6 +367,9 @@ class Command(BaseCommand):
 
 		self.stdout.write(f"{model_class.__name__} {id_}", ending="")
 
+		# For objects that we can't directly choose the ids of
+		# (and that will thus vary between stripe accounts)
+		# we fetch the id from a related object
 		if issubclass(model_class, djstripe.models.Account):
 			created, obj = self.get_or_create_stripe_account(
 				old_obj=old_obj, readonly_fields=readonly_fields
@@ -378,6 +388,14 @@ class Command(BaseCommand):
 			)
 		elif issubclass(model_class, djstripe.models.Charge):
 			created, obj = self.get_or_create_stripe_charge(
+				old_obj=old_obj, writable_fields=["metadata"]
+			)
+		elif issubclass(model_class, djstripe.models.PaymentIntent):
+			created, obj = self.get_or_create_stripe_payment_intent(
+				old_obj=old_obj, writable_fields=["metadata"]
+			)
+		elif issubclass(model_class, djstripe.models.PaymentMethod):
+			created, obj = self.get_or_create_stripe_payment_method(
 				old_obj=old_obj, writable_fields=["metadata"]
 			)
 		elif issubclass(model_class, djstripe.models.BalanceTransaction):
@@ -526,6 +544,63 @@ class Command(BaseCommand):
 				obj[k] = old_obj[k]
 
 		obj.save()
+
+		return created, obj
+
+	def get_or_create_stripe_payment_intent(self, old_obj, writable_fields):
+		invoice = djstripe.models.Invoice(id=old_obj["invoice"]).api_retrieve()
+		id_ = invoice["payment_intent"]
+
+		try:
+			obj = djstripe.models.PaymentIntent(id=id_).api_retrieve()
+			created = False
+
+			self.stdout.write(f"	found {id_}")
+		except InvalidRequestError:
+			assert False, "Expected to find payment_intent via invoice"
+
+		for k in writable_fields:
+			if isinstance(obj.get(k), dict):
+				# merge dicts (eg metadata)
+				obj[k].update(old_obj.get(k, {}))
+			else:
+				obj[k] = old_obj[k]
+
+		obj.save()
+
+		return created, obj
+
+	def get_or_create_stripe_payment_method(self, old_obj, writable_fields):
+		id_ = old_obj["id"]
+		customer_id = old_obj["customer"]
+		type_ = old_obj["type"]
+
+		try:
+			obj = djstripe.models.PaymentMethod(id=id_).api_retrieve()
+			created = False
+
+			self.stdout.write("	found")
+		except InvalidRequestError:
+			self.stdout.write("	creating")
+
+			obj = djstripe.models.PaymentMethod()._api_create(
+				type=type_, card={"token": "tok_visa"}
+			)
+
+			stripe.PaymentMethod.attach(
+				obj["id"], customer=customer_id, api_key=djstripe.settings.STRIPE_SECRET_KEY
+			)
+
+			for k in writable_fields:
+				if isinstance(obj.get(k), dict):
+					# merge dicts (eg metadata)
+					obj[k].update(old_obj.get(k, {}))
+				else:
+					obj[k] = old_obj[k]
+
+			obj.save()
+
+			created = True
 
 		return created, obj