+++ tests/test_invoice.py
@@ -13,8 +13,8 @@ from djstripe.settings import STRIPE_SECRET_KEY
 
 from . import (
 	FAKE_BALANCE_TRANSACTION, FAKE_CHARGE, FAKE_CUSTOMER, FAKE_INVOICE,
-	FAKE_INVOICEITEM_II, FAKE_PLAN, FAKE_PRODUCT, FAKE_SUBSCRIPTION,
-	FAKE_UPCOMING_INVOICE, IS_ASSERT_CALLED_AUTOSPEC_SUPPORTED,
+	FAKE_INVOICEITEM_II, FAKE_PAYMENT_INTENT_I, FAKE_PLAN, FAKE_PRODUCT,
+	FAKE_SUBSCRIPTION, FAKE_UPCOMING_INVOICE, IS_ASSERT_CALLED_AUTOSPEC_SUPPORTED,
 	IS_STATICMETHOD_AUTOSPEC_SUPPORTED, AssertStripeFksMixin, default_account
 )
 
@@ -42,10 +42,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_sync_from_stripe_data(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -66,6 +72,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -85,10 +94,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_retry_true(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -117,6 +132,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -136,10 +154,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_retry_false(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -165,6 +189,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -183,10 +210,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_status_paid(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -206,6 +239,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -224,10 +260,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_status_open(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -249,6 +291,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -267,10 +312,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_status_forgiven(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -292,6 +343,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -310,10 +364,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_status_forgiven_deprecated(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -345,10 +405,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_status_forgiven_default(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -379,10 +445,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_status_closed(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -404,6 +476,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -422,10 +497,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_status_closed_deprecated(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -458,10 +539,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_status_closed_default(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -496,10 +583,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 	)
 	@patch("stripe.Subscription.retrieve", autospec=IS_ASSERT_CALLED_AUTOSPEC_SUPPORTED)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_sync_no_subscription(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		plan_retrieve_mock,
@@ -531,6 +624,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
 				"djstripe.Invoice.subscription",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -549,10 +645,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_invoice_with_subscription_invoice_items(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -581,6 +683,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -599,10 +704,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_invoice_with_no_invoice_items(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -625,6 +736,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -643,10 +757,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_invoice_with_non_subscription_invoice_items(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -670,6 +790,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -688,10 +811,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_invoice_plan_from_invoice_items(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -713,6 +842,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -731,10 +863,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
 	def test_invoice_plan_from_subscription(
 		self,
 		product_retrieve_mock,
+		payment_intent_retrieve_mock,
 		charge_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
@@ -756,6 +894,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.dispute",
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -769,10 +910,16 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 		autospec=True,
 	)
 	@patch("stripe.Subscription.retrieve", autospec=True)
+	@patch(
+		"stripe.PaymentIntent.retrieve",
+		return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
+		autospec=True,
+	)
 	@patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
 	def test_invoice_without_plan(
 		self,
 		charge_retrieve_mock,
+		payment_intent_retrieve_mock,
 		subscription_retrieve_mock,
 		balance_transaction_retrieve_mock,
 		default_account_mock,
@@ -795,7 +942,10 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Charge.transfer",
 				"djstripe.Customer.coupon",
 				"djstripe.Invoice.subscription",
+				"djstripe.PaymentIntent.on_behalf_of",
+				"djstripe.PaymentIntent.payment_method",
 				"djstripe.Plan.product",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)