+++ tests/test_subscription.py
@@ -40,7 +40,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 			"{email} on {plan}".format(email=self.user.email, plan=str(subscription.plan)),
 		)
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -62,7 +68,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertTrue(self.customer.has_active_subscription())
 		self.assertTrue(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -82,7 +94,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertTrue(self.customer.has_active_subscription())
 		self.assertTrue(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -103,7 +121,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertFalse(self.customer.has_active_subscription())
 		self.assertFalse(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -134,7 +158,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertTrue(self.customer.has_active_subscription())
 		self.assertTrue(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -160,7 +190,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertFalse(self.customer.has_active_subscription())
 		self.assertFalse(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -193,7 +229,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertTrue(self.customer.has_active_subscription())
 		self.assertTrue(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -219,7 +261,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 
 		self.assertEqual(4, new_subscription.quantity)
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -243,7 +291,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 
 		self.assertEqual(Decimal(0.0), new_subscription.tax_percent)
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -270,7 +324,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 
 		self.assertEqual(FAKE_PLAN_II["id"], new_subscription.plan.id)
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 		self.assert_fks(new_plan, expected_blank_fks={})
 
@@ -313,7 +373,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertFalse(self.customer.has_active_subscription())
 		self.assertFalse(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -358,7 +424,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertTrue(self.customer.has_active_subscription())
 		self.assertTrue(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -397,7 +469,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertFalse(self.customer.has_active_subscription())
 		self.assertFalse(self.customer.has_any_active_subscription())
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", return_value=deepcopy(FAKE_PLAN), autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -437,7 +515,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		)
 		self.assertEqual(reactivated_subscription.cancel_at_period_end, False)
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("djstripe.models.Subscription._api_delete", autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -458,7 +542,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		subscription.cancel(at_period_end=False)
 		self.assertEqual(Subscription.objects.filter(status="canceled").count(), 1)
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("djstripe.models.Subscription._api_delete", autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -471,7 +561,13 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		with self.assertRaises(InvalidRequestError):
 			subscription.cancel(at_period_end=False)
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)
 
 	@patch("stripe.Plan.retrieve", autospec=True)
 	@patch("stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True)
@@ -503,6 +599,7 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 				"djstripe.Customer.coupon",
 				"djstripe.Customer.subscriber",
 				"djstripe.Subscription.plan",
+				"djstripe.Subscription.pending_setup_intent",
 			},
 		)
 
@@ -540,4 +637,10 @@ class SubscriptionTest(AssertStripeFksMixin, TestCase):
 		self.assertIsNone(item.quantity)
 		self.assertEqual(item.plan.id, FAKE_PLAN_METERED["id"])
 
-		self.assert_fks(subscription, expected_blank_fks={"djstripe.Customer.coupon"})
+		self.assert_fks(
+			subscription,
+			expected_blank_fks={
+				"djstripe.Customer.coupon",
+				"djstripe.Subscription.pending_setup_intent",
+			},
+		)